Player Identification Data

Obtain player identification data for various platforms.

Usage

get_player_data(season = NULL, pos = NULL)

Arguments

season

NFL season or year (int)

pos

Player position (str) such as "WR"

Value

Dataframe that lists player data for a specified season and position

Details

Use the function get_player_data to obtain player information for a defined season and position. Player information includes several IDs for platforms such as ESPN, Rotowire, ESPN, etc.

See also