QB Advanced Season Stats from Pro Football Reference (PFR)

Obtain QB advanced season stats from nflreadr, to acquire Pro Football Reference (PFR) data.

Usage

get_qb_pfr_advstats_season(seasons = NULL)

Arguments

seasons

NFL Season(s) where seasons = 2024

Value

Dataframe with QB advanced stats from Pro Football Reference (PFR) for a defined season

Details

get_qb_pfr_advstats_season is a function to obtain advanced stats from Pro Football Reference (PFR). The function utilizes the nflreadr function load_pfr_advstats to load player level season stats starting with the 2018 season. For quarterbacks, the stat_type is defined as pass. More information is available about QB PFR advanced stats from PFR passing data dictionary.

See also

Author

Nolan MacDonald

player_display_namePlayer name (e.g., Lamar Jackson)

pass_attemptsTotal pass attempts

throwawaysNumber of throwaways

spikesNumber of spikes

dropsNumber of throws dropped

drop_pctPercentage of dropped throws

bad_throwsNumber of bad throws

bad_throw_pctPercentage of bad throws

pocket_timeAverage time in pocket

times_blitzedNumber of times blitzed

times_hurriedNumber of times hurried

times_hitNumber of times hit

times_pressuredNumber of times pressured

pressure_pctPercent of time pressured

batted_ballsNumber of batted balls

on_tgt_throwsNumber of throws on target

on_tgt_pctPercent of throws on target

rpo_playsRun-pass-option (RPO) number of plays

rpo_yardsRun-pass-option (RPO) total yards

rpo_pass_attRun-pass-option (RPO) pass attempts

rpo_pass_yardsRun-pass-option (RPO) pass yards

rpo_rush_attRun-pass-option (RPO) rush attempts

rpo_rush_yardsRun-pass-option (RPO) rush yards

pa_pass_attPlay action pass attempts

pa_pass_yardsPlay action pass yards

pfr_player_id Pro Football Reference player ID (e.g., JackLa00)