RB Cumulative Season Advanced Stats

Obtain RB cumulative season stats from nflfastR, and using nflreadr to acquire Pro Football Reference (PFR) and Next Gen Stats (NGS) data

Usage

get_rb_adv_stats_season(
  pbp_db = "./data/pbp_db/pbp_db",
  pbp_db_tbl = "nflfastR_pbp",
  season = 2024
)

Arguments

pbp_db

Path to play-by-play database and database name

pbp_db_tbl

Play-by-play table name

season

NFL Season

Value

Dataframe with RB advanced statistics for a defined season

player

Player name

team

Player team in defined season

games

Total games played in defined season

fantasy_points

Total fantasy points in a standard format (0 points per reception)

fantasy_points_ppr

Total fantasy points in a full PPR format (1 point per reception)

ppg

Average fantasy points per game

Details

This is a function to obtain NFL advanced statistics for an entire season. The function utilizes nflfastR for play-by-play data to determine cumulative season statistics. The nflreadr package is also used to obtain advanced statistics from Pro Football Reference (PFR) and NFL Next Gen Stats (NGS). To use the function, you will need to save play-by-play data to a database using the nflfastR function, update_db(). For example, the database is saved by default as pbp_db with a table that is stored containing all play-by-play information called nflfastR_pbp. By defining the argument pbp_db (str) the user is specifying the path to the database as well as the name it was saved as. To load the table inside the database, pbp_db_tbl (str) can be defined if the naming is different, however, the default argument follows the naming scheme provided by the nflfastR package. The final requirement is the argument season (int) to indicate the NFL season that the user wants to obtain advanced statistics for. More information is available on update_db() in the nflfastR documentation.

See also

Author

Nolan MacDonald