RB Cumulative Season NFL Next Gen Stats (NGS)

Obtain RB cumulative season stats from nflreadr, to acquire Next Gen Stats (NGS) data.

Usage

get_rb_ngs_advstats_season(seasons = NULL)

Arguments

seasons

NFL Season(s) where seasons = 2024 or seasons = c(2023, 2024)

Value

Dataframe with RB Next Gen Stats for a defined season

Details

get_rb_ngs_advstats_season is a function to obtain NFL Next Gen Stats (NGS) for an entire season or multiple seasons. The function utilizes the nflreadr function load_nextgen_stats to load player level weekly stats starting with the 2016 season. For running backs, the stat_type is defined as rushing and is updated every night. NGS will only provide data for players above a minimum number of receptions. More information is available about NGS stats from NGS glossary.

See also

Author

Nolan MacDonald

player_idPlayer gsis id (e.g., 00-0038120)

player_display_namePlayer name (e.g., Breece Hall)

player_namePlayer shortened name (e.g., B.Hall)

positionPlayer position

rush_attemptsTotal rush attempts

rush_yardsTotal rush yards

avg_rush_yardsAverage rush yards per attempt

rush_touchdownsTotal rush touchdowns

efficiencyAverage efficiency (EFF)

percent_attempts_gte_eight_defenders Average % attempts with 8+ Defenders in the Box (8+D%)

avg_time_to_losAverage time behind line of scrimmage (TLOS)

expected_rush_yardsAverage expected rush yards

rush_yards_over_expectedAverage rush yards over expected

rush_yards_over_expected_per_att Average rush yards over expected per attempt

rush_pct_over_expectedAverage rush % over expected