RB Advanced Season Stats from Pro Football Reference (PFR)

Obtain RB advanced season stats from nflreadr, to acquire Pro Football Reference (PFR) data.

Usage

get_rb_pfr_advstats_season(seasons = NULL)

Format

A data frame with 16 variables that are described below.

player_display_name

Player name (e.g., Lamar Jackson)

age

position

g

Number of games played

gs

Number of games started

att

Total rush attempts

yds

Total rush yards

td

Total rush touchdowns

x1d

Rushing first downs

ybc

Rushing yards before contact

ybc_att

Rushing yards before contact per rushing attempt

yac

Rushing yards after contact

yac_att

Rushing yards after contact per attempt

brk_tkl

Broken tackles on rushes

att_br

Rush attempts per broken tackle

pfr_player_id

Pro Football Reference player ID (e.g., JackLa00)

Arguments

seasons

NFL Season(s) where seasons = 2024

Value

Dataframe with RB advanced stats from Pro Football Reference (PFR) for a defined season

Details

get_rb_pfr_advstats_season is a function to obtain advanced stats from Pro Football Reference (PFR). The function utilizes the nflreadr function load_pfr_advstats to load player level season stats starting with the 2018 season. For running backs, the stat_type is defined as rush. More information is available about RB PFR advanced stats from seasons on the PFR site such as PFR 2024 NFL Advanced Rushing.

See also

Author

Nolan MacDonald