Helper Function - Get Scoring Columns

Obtain the valid columns to use in dplyr::select for a user-defined scoring format.

Usage

get_scoring_columns(scoring = NULL)

Arguments

scoring

Fantasy scoring format (str). Can be defined as standard scoring, half PPR scoring, full PPR scoring, 4PT PASS TD or 6PT PASS TD by using: "std_4pt_td", "half_ppr_4pt_td", "ppr_4pt_td", "std_6pt_td", "half_ppr_6pt_td", "ppr_6pt_td"

Value

Defined column names for fantasy points and points per game (PPG)

Details

Depending on the user-defined argument when creating gt tables for fantasy stats, the scoring format columns can be included while neglecting all other calculated fantasy scoring formats.

See also

Author

Nolan MacDonald