Player Snap Share

Obtain player snap shares provided by Pro Football Reference (PFR), starting with the 2012 season.

Usage

get_snap_share(season = NULL, pos = NULL)

Arguments

pos

Player position (str)

seasons

NFL Season(s) (int) where seasons = 2024 or seasons = c(2023, 2024)

Value

Obtain player snap share percentage for positional group

Details

This function loads game level snap counts by PFR starting with the 2012 season. Snap count data is then utilized to filter by season and player position. Additional data wrangling is performed to format player names in order to merge easily with play-by-play data. More information is available about snap counts from Load Snap Counts from PFR.

See also

Author

Nolan MacDonald

playerPlayer name (e.g., Darnell Mooney)

positionPlayer name (e.g., WR)

teamPlayer team

pfr_player_idPro Football Reference (PFR) player ID

snapsNumber of snaps

snap_shareFraction of total snaps

snap_pctPercentage of total snaps