TE Advanced Season Stats from Pro Football Reference (PFR)

Obtain TE advanced season stats from nflreadr, to acquire Pro Football Reference (PFR) data.

Usage

get_te_pfr_advstats_season(seasons = NULL)

Arguments

seasons

NFL Season(s) where seasons = 2024

Value

Dataframe with TE advanced stats from Pro Football Reference (PFR) for a defined season

Details

get_te_pfr_advstats_season is a function to obtain advanced stats from Pro Football Reference (PFR). The function utilizes the nflreadr function load_pfr_advstats to load player level season stats starting with the 2018 season. For receivers, the stat_type is defined as rec. More information is available about TE PFR advanced stats from seasons on the PFR site such as PFR 2024 NFL Advanced Receiving.

See also

Author

Nolan MacDonald

player_display_namePlayer name (e.g., George Kittle)

agePlayer age

positionPlayer position

gNumber of games played

gsNumber of games started

tgtTotal targets

recTotal receptions

ydsTotal receiving yards

tdTotal receiving touchdowns

x1dFirst downs receiving

ybc Total yards passes traveled in the air before being caught or yards before catch (YBC)

ybc_rYards before catch per reception

yacYards after catch (YAC)

yac_rYards after catch (YAC) per reception

adot Average depth of target (ADOT) when targeted, whether completed or not.

brk_tklNumber of broken tackles

rec_brReceptions per broken tackle

dropNumber of dropped passes

drop_percentDropped pass percentage when targeted

int_tgtInterceptions on passes where targeted

ratPasser rating on passes when targeted

pfr_player_id Pro Football Reference player ID (e.g., KittGe00)