Helper - Validate Play-by-Play Database Path and Table Arguments

Validation helper to determine if the path to the play-by-play database, pbp_db, and the database table name, pbp_db_tbl, are defined as strings.

Usage

validate_pbp_db(pbp_db, pbp_db_tbl)

Arguments

pbp_db_tbl

Play-by-Play database table name (str)

pbp_dp

Play-by-Play database path (str)

Value

Validate if pbp_db and pbp_db_tbl are strings

Details

The function validate_pbp_db can be utilized to validate whether the arguments to load a database in get_pbp_data are valid and defined as strings.

See also

Author

Nolan MacDonald